Further refactor parameters to live in a module
