+++ lepo/apidef/parameter/base.py
@@ -1,18 +1,7 @@
-import json
-
 import jsonschema
-from django.utils.encoding import force_text
 
-from lepo.excs import InvalidBodyContent, InvalidBodyFormat
 from lepo.parameter_utils import cast_primitive_value
 from lepo.utils import maybe_resolve
-
-COLLECTION_FORMAT_SPLITTERS = {
-    'csv': lambda value: force_text(value).split(','),
-    'ssv': lambda value: force_text(value).split(' '),
-    'tsv': lambda value: force_text(value).split('\t'),
-    'pipes': lambda value: force_text(value).split('|'),
-}
 
 OPENAPI_JSONSCHEMA_VALIDATION_KEYS = (
     'maximum', 'exclusiveMaximum',
@@ -28,6 +17,9 @@ OPENAPI_JSONSCHEMA_VALIDATION_KEYS = (
 class BaseParameter:
     def __init__(self, data):
         self.data = data
+
+    def __repr__(self):
+        return '<%s (%r)>' % (self.__class__.__name__, self.data)
 
     @property
     def location(self):
@@ -56,10 +48,6 @@ class BaseParameter:
     @property
     def default(self):
         return self.data.get('default')
-
-    @property
-    def collection_format(self):
-        return self.data.get('collectionFormat')
 
     @property
     def validation_keys(self):
@@ -96,12 +84,7 @@ class BaseParameter:
         return [items_param.cast(api, item) for item in value]
 
     def arrayfy_value(self, value):
-        collection_format = self.collection_format or 'csv'
-        splitter = COLLECTION_FORMAT_SPLITTERS.get(collection_format)
-        if not splitter:
-            raise NotImplementedError('unsupported collection format in %r' % self)
-        value = splitter(value)
-        return value
+        raise NotImplementedError('...')
 
     def cast(self, api, value):
         if self.type == 'array':
@@ -122,13 +105,13 @@ class BaseParameter:
         raise NotImplementedError('...')
 
 
-class Parameter(BaseParameter):
+class BaseTopParameter(BaseParameter):
     """
     Top-level Parameter, such as in an operation
     """
 
     def __init__(self, data, operation=None):
-        super(Parameter, self).__init__(data)
+        super(BaseTopParameter, self).__init__(data)
         self.operation = operation
 
     @property
@@ -147,53 +130,3 @@ class Parameter(BaseParameter):
             return request.META['HTTP_%s' % self.name.upper().replace('-', '_')]
 
         raise NotImplementedError('unsupported `in` value %r in %r' % (self.location, self))  # pragma: no cover
-
-
-class Swagger2Parameter(Parameter):
-
-    def get_value(self, request, view_kwargs):
-        if self.location == 'body':
-            return self.read_body(request)
-
-        if self.location == 'formData' and self.type == 'file':
-            return request.FILES[self.name]
-
-        if self.location in ('query', 'formData'):
-            source = (request.POST if self.location == 'formData' else request.GET)
-            print(self.name, self.data)
-            if self.type == 'array' and self.collection_format == 'multi':
-                return source.getlist(self.name)
-            else:
-                return source[self.name]
-
-        return super().get_value(request, view_kwargs)
-
-    def read_body(self, request):
-        consumes = request.api_info.operation.consumes
-        if request.content_type not in consumes:
-            raise InvalidBodyFormat('Content-type %s is not supported (%r are)' % (
-                request.content_type,
-                consumes,
-            ))
-
-        return read_body(request, self)
-
-
-def read_body(request, parameter=None):
-    if parameter:
-        if parameter.type == 'binary':
-            return request.body.read()
-    try:
-        if request.content_type == 'application/json':
-            return json.loads(request.body.decode(request.content_params.get('charset', 'UTF-8')))
-        elif request.content_type == 'text/plain':
-            return request.body.decode(request.content_params.get('charset', 'UTF-8'))
-        if request.content_type == 'multipart/form-data':
-            # TODO: this definitely doesn't handle multiple values for the same key correctly
-            data = dict()
-            data.update(request.POST.items())
-            data.update(request.FILES.items())
-            return data
-    except Exception as exc:
-        raise InvalidBodyContent('Unable to parse this body as %s' % request.content_type) from exc
-    raise NotImplementedError('No idea how to parse content-type %s' % request.content_type)  # pragma: no cover
